+++ metaworld/envs/mujoco/sawyer_xyz/v1/sawyer_hammer.py
@@ -35,7 +35,10 @@ class SawyerHammerEnv(SawyerXYZEnv):
         self.liftThresh = liftThresh
 
         self._random_reset_space = Box(np.array(obj_low), np.array(obj_high))
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
 
     @property
     def model_name(self):
