+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_door_close_v2.py
@@ -41,11 +41,14 @@ class SawyerDoorCloseEnvV2(SawyerXYZEnv):
         self.door_qpos_adr = self.model.joint("doorjoint").qposadr.item()
         self.door_qvel_adr = self.model.joint("doorjoint").dofadr.item()
 
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
 
         self._random_reset_space = Box(
-            np.array(obj_low),
-            np.array(obj_high),
+            np.array(obj_low, dtype=np.float32),
+            np.array(obj_high, dtype=np.float32),
         )
 
     @property
