+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_drawer_close_v2.py
@@ -46,10 +46,13 @@ class SawyerDrawerCloseEnvV2(SawyerXYZEnv):
         goal_high = self.hand_high
 
         self._random_reset_space = Box(
-            np.array(obj_low),
-            np.array(obj_high),
+            np.array(obj_low, dtype=np.float32),
+            np.array(obj_high, dtype=np.float32),
         )
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
 
         self.maxDist = 0.15
         self.target_reward = 1000 * self.maxDist + 1000 * 2
