+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_sweep_v2.py
@@ -43,10 +43,13 @@ class SawyerSweepEnvV2(SawyerXYZEnv):
         self.init_puck_z = init_puck_z
 
         self._random_reset_space = Box(
-            np.array(obj_low),
-            np.array(obj_high),
+            np.array(obj_low, dtype=np.float32),
+            np.array(obj_high, dtype=np.float32),
         )
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
 
     @property
     def model_name(self):
